% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2compareGroups-package.R
\docType{package}
\name{add2compareGroups-package}
\alias{add2compareGroups-package}
\title{add2compareGroups: Add adjusted OR/HR to compareGroups tables}
\description{
The add2compareGroups package has been developed to extend the univariate descriptive tables provided by the compareGroups package (Subirana, Sanz, and Vila 2014). It allows users to add adjusted Odds Ratios (OR) and Hazard Ratios (HR) directly into those tables. This extension is especially useful in real-world data (RWD) settings where raw and adjusted association measures need to be presented in the same report. Moreover, the package provides the ability to generate Forest Plot graphs for both raw and adjusted OR/HR, facilitating their visual comparison and interpretation.
}
\author{
\strong{Maintainer}: Mireia Roca Loureiro \email{mireia.roca.data@gmail.com}

}
\keyword{internal}
