% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_adjusted_ratio.R
\name{add_adjusted_ratio}
\alias{add_adjusted_ratio}
\title{Add adjusted OR/HR columns to compareGroups objects}
\usage{
add_adjusted_ratio(
  compare_object,
  adjust_common = NULL,
  adjust = NULL,
  show.ratio = TRUE,
  show.p.ratio = show.ratio,
  show.p.overall = TRUE,
  show.n = FALSE
)
}
\arguments{
\item{compare_object}{An object of class compareGroups, createTable or descrTable that contains the original comparisons.}

\item{adjust_common}{A character list specifying a common set of covariables used to adjust all variables included in compare_object.}

\item{adjust}{A names list where each element corresponds to a variable in compare_object, and its value is a character list containing the covariates used to adjust that specific variable.}

\item{show.ratio}{Logical. If TRUE, adjusted OR/HR are shown in the table.}

\item{show.p.ratio}{Logical. If TRUE, p-values corresponding to each adjusted OR/HR is diplayed in the table.}

\item{show.p.overall}{Logical. If TRUE, the overall p-values for group significance ('p.overall' column) is displayed in the table.}

\item{show.n}{Logical. If TRUE, the number of observations used to compute each OR/HR estimate is displayed in the table.}
}
\value{
An updated object of class add_adjusted_ratio and createTable, generated from the input compare_object, that now includes additional columns with the adjusted OR/HR values and their associated statistics.
}
\description{
This function adds new columns with adjusted Odds Ratios (OR) or Hazard Ratios (HR), along with their respective statistical parameters (such as p-values and p-overall), to an object from compareGroups family. Users can specify different adjustment variables for each descripted variable in the comparison or define a common set of covariables to be applied across all variables.
}
\examples{
\dontrun{

library("compareGroups")
library("add2compareGroups")

data(lung_patients)

# Create inicial compareGroups objects
cg_death <- compareGroups(
  DEATH ~ GENDER + AGE + SMOKING + PACK_YEARS + RECURRENCE,
  data = lung_patients
) # Create a basic compareGroups

# Transform it to a createTable for better late interpretation
ct_death <- createTable(cg_death, show.ratio = TRUE)

# Common adjusted variables
OR_common <- add_adjusted_ratio(ct_death,
  adjust_common = list("AGE", "GENDER")
)
export2md(OR_common)

# Individual adjusted variables
OR_individual <- add_adjusted_ratio(ct_death,
  adjust_common = list("AGE", "GENDER"),
   adjust = list("SMOKING" = list("PACK_YEARS", "PROGRESSION"), "RECURRENCE" = list("SURGERY")),
   show.n = TRUE
)
export2md(OR_individual)

}
}
