% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_forest_plot.R
\name{ratio_forest_plot}
\alias{ratio_forest_plot}
\title{Create unadjusted and adjusted OR/HR forest plot}
\usage{
ratio_forest_plot(compare_object, hline_forest_plot = 1, log_scale = FALSE)
}
\arguments{
\item{compare_object}{An object of class compareGroups family or add_adjusted_ratio with de OR/HR calculated for plotting.}

\item{hline_forest_plot}{numeric value indicating the position of the horizontal reference line in the forest plot. Default is 1, representing the null effect. If you do not want to display it, you can set the value to NULL.}

\item{log_scale}{Logical. If TRUE, the OR/HR will be displayed on a logarithmic scale in the forest plot.}
}
\value{
A forest_plot with the OR/HR according to the input object.
}
\description{
This function creates forest plots of OR/HR. It can make unadjusted if the entrance object is from compareGroups family, or adjusted if it is an add_adjusted_ratio object.
}
\examples{
\dontrun{

library("compareGroups")
library("add2compareGroups")

data(lung_patients)

# Create inicial compareGroups objects
cg_death <- compareGroups(
  DEATH ~ GENDER + AGE + SMOKING + PACK_YEARS + RECURRENCE,
  data = lung_patients
) # Create a basic compareGroups

# Transform it to a createTable for better late interpretation
ct_death <- createTable(cg_death, show.ratio = TRUE)

# Common adjusted variables
OR_common <- add_adjusted_ratio(ct_death, adjust_common = list("AGE", "GENDER"))
export2md(OR_common)

# Individual adjusted variables
OR_individual <- add_adjusted_ratio(ct_death,
  adjust_common = list("AGE", "GENDER"),
  adjust = list("SMOKING" = list("PACK_YEARS", "PROGRESSION"), "RECURRENCE" = list("SURGERY")),
  show.n = TRUE
)
export2md(OR_individual)

# Create the forest_plots
ratio_forest_plot(ct_death)
ratio_forest_plot(ct_death, log_scale = TRUE)
ratio_forest_plot(OR_individual, hline_forest_plot = NULL)
ratio_forest_plot(OR_individual, log_scale = TRUE)

}
}
